+++ crates/plugin/src/plugin_registry.rs
@@ -157,7 +157,7 @@ impl<T: Plugin> PluginRegistry<T> {
                     self.host_data.clone(),
                     HostData {
                         cache_dir: self.host_data.moon_env.cache_dir.clone(),
-                        http_client: self.loader.get_client()?.clone(),
+                        http_client: self.loader.get_http_client()?.clone(),
                         virtual_paths: self.virtual_paths.clone(),
                         working_dir: self.host_data.moon_env.working_dir.clone(),
                     },