+++ legacy/node/tool/src/node_tool.rs
@@ -306,7 +306,9 @@ impl Tool for NodeTool {
     }
 
     async fn teardown(&mut self) -> miette::Result<()> {
-        self.tool.teardown().await?;
+        if let Some(version) = self.config.version.clone() {
+            self.tool.teardown(&ToolSpec::new(version)).await?;
+        }
 
         Ok(())
     }