+++ legacy/node/tool/src/pnpm_tool.rs
@@ -127,7 +127,9 @@ impl Tool for PnpmTool {
     }
 
     async fn teardown(&mut self) -> miette::Result<()> {
-        self.tool.teardown().await?;
+        if let Some(version) = self.config.version.clone() {
+            self.tool.teardown(&ToolSpec::new(version)).await?;
+        }
 
         Ok(())
     }