+++ legacy/python/tool/src/python_tool.rs
@@ -294,7 +294,9 @@ impl Tool for PythonTool {
     }
 
     async fn teardown(&mut self) -> miette::Result<()> {
-        self.tool.teardown().await?;
+        if let Some(version) = self.config.version.clone() {
+            self.tool.teardown(&ToolSpec::new(version)).await?;
+        }
 
         Ok(())
     }